refactor(e2e): clarify console session ownership
