+++ e2e/scripts/seed.ts
@@ -4,7 +4,7 @@ import path from 'node:path'
 import { chromium } from '@playwright/test'
 import { createAgentV2SeedTasks } from '../features/agent-v2/support/seed'
 import { ensureAuthenticatedState } from '../fixtures/auth'
-import { createStandaloneConsoleSession } from '../support/api/console-context'
+import { createStandaloneConsoleSession } from '../support/api/console-session'
 import { startLoggedProcess, stopManagedProcess, waitForUrl } from '../support/process'
 import { runSeedTasks, writeSeedReport } from '../support/seed'
 import { startWebServer, stopWebServer } from '../support/web-server'
